+++ ui/src/organizations/components/VariableList.tsx
@@ -7,7 +7,7 @@ import VariableRow from 'src/organizations/components/VariableRow'
 import UpdateVariableOverlay from 'src/organizations/components/UpdateVariableOverlay'
 
 // Types
-import {Variable} from '@influxdata/influx'
+import {IVariable as Variable} from '@influxdata/influx'
 import {OverlayState} from 'src/types'
 
 interface Props {
@@ -15,6 +15,7 @@ interface Props {
   emptyState: JSX.Element
   onDeleteVariable: (variable: Variable) => void
   onUpdateVariable: (variable: Variable) => void
+  onFilterChange: (searchTerm: string) => void
 }
 
 interface State {
@@ -38,6 +39,7 @@ class VariableList extends PureComponent<Props, State> {
       variables,
       onDeleteVariable,
       onUpdateVariable,
+      onFilterChange,
     } = this.props
 
     return (
@@ -55,6 +57,7 @@ class VariableList extends PureComponent<Props, State> {
                 onDeleteVariable={onDeleteVariable}
                 onUpdateVariableName={onUpdateVariable}
                 onEditVariable={this.handleStartEdit}
+                onFilterChange={onFilterChange}
               />
             ))}
           </IndexList.Body>