+++ ui/src/organizations/components/VariableRow.tsx
@@ -1,39 +1,77 @@
 // Libraries
 import React, {PureComponent} from 'react'
+import {connect} from 'react-redux'
 import {withRouter, WithRouterProps} from 'react-router'
 
 // Components
-import {IndexList, Alignment, Context, IconFont} from 'src/clockface'
+import {
+  IndexList,
+  Alignment,
+  Context,
+  IconFont,
+  Stack,
+  ComponentSpacer,
+} from 'src/clockface'
 import {ComponentColor} from '@influxdata/clockface'
+import InlineLabels from 'src/shared/components/inlineLabels/InlineLabels'
 
 // Types
-import {Variable} from '@influxdata/influx'
+import {IVariable as Variable, ILabel} from '@influxdata/influx'
 import EditableName from 'src/shared/components/EditableName'
+import {AppState} from 'src/types'
+
+// Selectors
+import {viewableLabels} from 'src/labels/selectors'
+
+// Actions
+import {
+  addVariableLabelsAsync,
+  removeVariableLabelsAsync,
+} from 'src/variables/actions'
+import {createLabel as createLabelAsync} from 'src/labels/actions'
 
 interface OwnProps {
   variable: Variable
   onDeleteVariable: (variable: Variable) => void
   onUpdateVariableName: (variable: Partial<Variable>) => void
   onEditVariable: (variable: Variable) => void
+  onFilterChange: (searchTerm: string) => void
 }
 
-type Props = OwnProps & WithRouterProps
+interface StateProps {
+  labels: ILabel[]
+}
 
-class VariableRow extends PureComponent<Props> {
+interface DispatchProps {
+  onAddVariableLabels: typeof addVariableLabelsAsync
+  onRemoveVariableLabels: typeof removeVariableLabelsAsync
+  onCreateLabel: typeof createLabelAsync
+}
+
+type Props = OwnProps & DispatchProps & StateProps
+
+class VariableRow extends PureComponent<Props & WithRouterProps> {
   public render() {
     const {variable, onDeleteVariable} = this.props
 
     return (
       <IndexList.Row testID="variable-row">
         <IndexList.Cell alignment={Alignment.Left}>
-          <EditableName
-            onUpdate={this.handleUpdateVariableName}
-            name={variable.name}
-            noNameString="NAME THIS VARIABLE"
-            onEditName={this.handleEditVariable}
+          <ComponentSpacer
+            stackChildren={Stack.Rows}
+            align={Alignment.Left}
+            stretchToFitWidth={true}
           >
-            {variable.name}
-          </EditableName>
+            <EditableName
+              onUpdate={this.handleUpdateVariableName}
+              name={variable.name}
+              noNameString="NAME THIS VARIABLE"
+              onEditName={this.handleEditVariable}
+            >
+              {variable.name}
+            </EditableName>
+            {this.labels}
+          </ComponentSpacer>
         </IndexList.Cell>
         <IndexList.Cell alignment={Alignment.Left}>Query</IndexList.Cell>
         <IndexList.Cell revealOnHover={true} alignment={Alignment.Right}>
@@ -59,6 +97,43 @@ class VariableRow extends PureComponent<Props> {
     )
   }
 
+  private get labels(): JSX.Element {
+    const {variable, labels, onFilterChange} = this.props
+    const collectorLabels = viewableLabels(variable.labels)
+
+    return (
+      <InlineLabels
+        selectedLabels={collectorLabels}
+        labels={labels}
+        onFilterChange={onFilterChange}
+        onAddLabel={this.handleAddLabel}
+        onRemoveLabel={this.handleRemoveLabel}
+        onCreateLabel={this.handleCreateLabel}
+      />
+    )
+  }
+
+  private handleAddLabel = (label: ILabel): void => {
+    const {variable, onAddVariableLabels} = this.props
+
+    onAddVariableLabels(variable.id, [label])
+  }
+
+  private handleRemoveLabel = (label: ILabel): void => {
+    const {variable, onRemoveVariableLabels} = this.props
+
+    onRemoveVariableLabels(variable.id, [label])
+  }
+
+  private handleCreateLabel = async (label: ILabel): Promise<void> => {
+    try {
+      const {name, properties} = label
+      await this.props.onCreateLabel(name, properties)
+    } catch (err) {
+      throw err
+    }
+  }
+
   private handleExport = () => {
     const {
       router,
@@ -79,4 +154,19 @@ class VariableRow extends PureComponent<Props> {
   }
 }
 
-export default withRouter<OwnProps>(VariableRow)
+const mstp = ({labels}: AppState): StateProps => {
+  return {
+    labels: viewableLabels(labels.list),
+  }
+}
+
+const mdtp: DispatchProps = {
+  onCreateLabel: createLabelAsync,
+  onAddVariableLabels: addVariableLabelsAsync,
+  onRemoveVariableLabels: removeVariableLabelsAsync,
+}
+
+export default connect<StateProps, DispatchProps, OwnProps>(
+  mstp,
+  mdtp
+)(withRouter<Props>(VariableRow))