+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 1. [13024](https://github.com/influxdata/influxdb/pull/13024): Add the ability to edit token's description
 1. [13078](https://github.com/influxdata/influxdb/pull/13078): Add the option to create a Dashboard from a Template.
+1. [13161](https://github.com/influxdata/influxdb/pull/13161): Add the ability to add labels on variables
 
 ### Bug Fixes